move registry to its own package
